+++ structurizr-client/src/main/java/com/structurizr/view/ThemeUtils.java
@@ -6,17 +6,22 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.structurizr.Workspace;
 import com.structurizr.io.WorkspaceWriterException;
+import com.structurizr.model.Relationship;
+import com.structurizr.util.ImageUtils;
 import com.structurizr.util.StringUtils;
+import com.structurizr.util.Url;
 import org.apache.hc.client5.http.classic.methods.HttpGet;
 import org.apache.hc.client5.http.config.ConnectionConfig;
 import org.apache.hc.client5.http.impl.classic.CloseableHttpClient;
 import org.apache.hc.client5.http.impl.classic.CloseableHttpResponse;
 import org.apache.hc.client5.http.impl.classic.HttpClientBuilder;
+import org.apache.hc.client5.http.impl.classic.HttpClients;
 import org.apache.hc.client5.http.impl.io.BasicHttpClientConnectionManager;
 import org.apache.hc.core5.http.io.entity.EntityUtils;
 
 import java.io.*;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
 import java.util.concurrent.TimeUnit;
 
 /**
@@ -65,7 +70,7 @@ public final class ThemeUtils {
     }
 
     /**
-     * Loads (and inlines) the element and relationship styles from the themes defined in the workspace, into the workspace itself.
+     * Loads the element and relationship styles from the themes defined in the workspace, into the workspace itself.
      * This implementation simply copies the styles from all themes into the workspace.
      * This uses a default timeout value of 10000ms.
      *
@@ -77,7 +82,7 @@ public final class ThemeUtils {
     }
 
     /**
-     * Loads (and inlines) the element and relationship styles from the themes defined in the workspace, into the workspace itself.
+     * Loads the element and relationship styles from the themes defined in the workspace, into the workspace itself.
      * This implementation simply copies the styles from all themes into the workspace.
      *
      * @param workspace                 a Workspace object
@@ -85,32 +90,11 @@ public final class ThemeUtils {
      * @throws Exception    if something goes wrong
      */
     public static void loadThemes(Workspace workspace, int timeoutInMilliseconds) throws Exception {
-        for (String url : workspace.getViews().getConfiguration().getThemes()) {
-            ConnectionConfig connectionConfig = ConnectionConfig.custom()
-                    .setConnectTimeout(timeoutInMilliseconds, TimeUnit.MILLISECONDS)
-                    .setSocketTimeout(timeoutInMilliseconds, TimeUnit.MILLISECONDS)
-                    .build();
-
-            BasicHttpClientConnectionManager cm = new BasicHttpClientConnectionManager();
-            cm.setConnectionConfig(connectionConfig);
-
-            CloseableHttpClient httpClient = HttpClientBuilder.create()
-                    .useSystemProperties()
-                    .setConnectionManager(cm)
-                    .build();
-
-            HttpGet httpGet = new HttpGet(url);
-
-            CloseableHttpResponse response = httpClient.execute(httpGet);
-            if (response.getCode() == HTTP_OK_STATUS) {
-                String json = EntityUtils.toString(response.getEntity());
-
-                ObjectMapper objectMapper = new ObjectMapper();
-                objectMapper.enable(DeserializationFeature.ACCEPT_EMPTY_STRING_AS_NULL_OBJECT);
-                objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-
-                Theme theme = objectMapper.readValue(json, Theme.class);
-                String baseUrl = url.substring(0, url.lastIndexOf('/') + 1);
+        for (String themeLocation : workspace.getViews().getConfiguration().getThemes()) {
+            if (Url.isUrl(themeLocation)) {
+                String json = loadFrom(themeLocation, timeoutInMilliseconds);
+                Theme theme = fromJson(json);
+                String baseUrl = themeLocation.substring(0, themeLocation.lastIndexOf('/') + 1);
 
                 for (ElementStyle elementStyle : theme.getElements()) {
                     String icon = elementStyle.getIcon();
@@ -128,9 +112,69 @@ public final class ThemeUtils {
 
                 workspace.getViews().getConfiguration().getStyles().addStylesFromTheme(theme);
             }
-
-            httpClient.close();
         }
+    }
+
+    /**
+     * Inlines the element and relationship styles from the specified file, adding the styles into the workspace
+     * and overriding any properties already set.
+     *
+     * @param workspace     the Workspace to load the theme into
+     * @param file          a File object representing a theme (a JSON file)
+     * @throws Exception    if something goes wrong
+     */
+    public static void inlineTheme(Workspace workspace, File file) throws Exception {
+        String json = Files.readString(file.toPath());
+        Theme theme = fromJson(json);
+
+        for (ElementStyle elementStyle : theme.getElements()) {
+            String icon = elementStyle.getIcon();
+            if (!StringUtils.isNullOrEmpty(icon)) {
+                if (icon.startsWith("http")) {
+                    // okay, image served over HTTP
+                } else if (icon.startsWith("data:image")) {
+                    // also okay, data URI
+                } else {
+                    // convert the relative icon filename into a data URI
+                    elementStyle.setIcon(ImageUtils.getImageAsDataUri(new File(file.getParentFile(), icon)));
+                }
+            }
+        }
+
+        workspace.getViews().getConfiguration().getStyles().inlineTheme(theme);
+    }
+
+    private static String loadFrom(String url, int timeoutInMilliseconds) throws Exception {
+        ConnectionConfig connectionConfig = ConnectionConfig.custom()
+                .setConnectTimeout(timeoutInMilliseconds, TimeUnit.MILLISECONDS)
+                .setSocketTimeout(timeoutInMilliseconds, TimeUnit.MILLISECONDS)
+                .build();
+
+        BasicHttpClientConnectionManager cm = new BasicHttpClientConnectionManager();
+        cm.setConnectionConfig(connectionConfig);
+
+        try (CloseableHttpClient httpClient = HttpClientBuilder.create()
+                .useSystemProperties()
+                .setConnectionManager(cm)
+                .build()) {
+
+            HttpGet httpGet = new HttpGet(url);
+
+            CloseableHttpResponse response = httpClient.execute(httpGet);
+            if (response.getCode() == HTTP_OK_STATUS) {
+                return EntityUtils.toString(response.getEntity());
+            }
+        }
+
+        return "";
+    }
+
+    private static Theme fromJson(String json) throws Exception {
+        ObjectMapper objectMapper = new ObjectMapper();
+        objectMapper.enable(DeserializationFeature.ACCEPT_EMPTY_STRING_AS_NULL_OBJECT);
+        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+
+        return objectMapper.readValue(json, Theme.class);
     }
 
     private static void write(Workspace workspace, Writer writer) throws Exception {